move definitions and settings to separate file
